Ageファイルで値とメソッドの定義
